Remove underscores
The undersores are not visible when results are displayed and commands are typed in terminal. This confuses the user.
For instance "ID_num" will be confused with "ID" and " num", which in first case is a single attribute formed by joining
"ID" and "num" with underscore and in second case "ID" and "num" are two separate attributes